US12 client.rb WIP
