+++ src/Portergeist/Server.php
@@ -29,11 +29,27 @@ class Server {
   /** @var  bool */
   protected $started;
 
+  /** @var Server */
+  private static $instance = null;
+
+  /**
+   * Singleton pattern because a server is a server you can only create it ONCE
+   * @param int $fixedPort
+   * @param int $timeout
+   * @return Server
+   */
+  public static function getInstance($fixedPort = null, $timeout = null) {
+    if (null === self::$instance) {
+      self::$instance = new self($fixedPort, $timeout);
+    }
+    return self::$instance;
+  }
+
   /**
    * @param int $fixedPort
    * @param int $timeout
    */
-  public function __construct($fixedPort = null, $timeout = null) {
+  private function __construct($fixedPort = null, $timeout = null) {
     $this->fixedPort = ($fixedPort === null) ? Server::DEFAULT_PORT : $fixedPort;
     $this->timeout = ($timeout === null) ? Server::BIND_TIMEOUT : $timeout;
     $this->wsClient = null;
